requirement.txt
fastapi
chromadb==0.3.21
sentence-transformers
uvicorn
pdfplumber
python-docx
pypika==0.48.0
